Change private member from parent model class to protected; Now compilation works; Need to add assignment overloading to Model class

+++ include/utilities/parameter_reader.hpp
@@ -20,7 +20,7 @@ namespace params
 {
 	extern std::string file_name;
 	extern int horizon;
-	extern int epsilon;
+	extern double epsilon;
 	extern int start_state;
 	extern std::vector<int> start_belief;
 	extern std::string solver_type;
