+++ src/parameter_reader.cpp
@@ -7,7 +7,7 @@ namespace params
 {
 	std::string file_name;
 	int horizon;
-	int epsilon;
+	double epsilon;
 	int start_state;
 	std::vector<int> start_belief;
 	std::string solver_type;
@@ -31,7 +31,7 @@ void read_parameters(int ac, char* av[])
 	"\n'long_option_name = value'\nand comment lines that begin with '#'." )
 	("file_name",po::value<std::string>(&params::file_name),"Location and name of the file containing the model information in cassandra format.") 
 	("horizon",po::value<int>(&params::horizon),"Horizon of MPC POMDP to solve the optimal control distribution (horizon >= 1).") 
-	("epsilon",po::value<int>(&params::epsilon),"Chance constraints rate (0 <= epsilon <= 1).") 
+	("epsilon",po::value<double>(&params::epsilon),"Chance constraints rate (0 <= epsilon <= 1).") 
 	("start_state",po::value<int>(&params::start_state),"Start state index of the system.") 
 	("start_belief",po::value<std::string>(),"The start belief index of the system. Belief is evenly distributed. Input is in the format of \"0 1\"") 
 	("solver_type",po::value<std::string>(&params::solver_type),"Type of solver to use (ipopt or nlopt).")
